refactor(bazel-module): consolidate kv parsing logic
This causes some spurious test changes, where more of the input can be
parsed now.
