+++ lib/modules/manager/bazel-module/parser/module.ts
@@ -1,9 +1,8 @@
 import { query as q } from 'good-enough-parser';
 import { regEx } from '../../../../util/regex';
 import type { Ctx } from '../context';
-import * as starlark from '../starlark';
+import { kvParams } from './common';
 
-const booleanValuesRegex = regEx(`^${starlark.booleanStringValues.join('|')}$`);
 const supportedRules = [
   'archive_override',
   'bazel_dep',
@@ -12,19 +11,6 @@ const supportedRules = [
   'single_version_override',
 ];
 const supportedRulesRegex = regEx(`^${supportedRules.join('|')}$`);
-
-/**
- * Matches key-value pairs:
- * - `name = "foobar"`
- * - `name = True`
- **/
-const kvParams = q
-  .sym<Ctx>((ctx, token) => ctx.startAttribute(token.value))
-  .op('=')
-  .alt(
-    q.str((ctx, token) => ctx.addString(token.value)),
-    q.sym<Ctx>(booleanValuesRegex, (ctx, token) => ctx.addBoolean(token.value)),
-  );
 
 export const moduleRules = q
   .sym<Ctx>(supportedRulesRegex, (ctx, token) => ctx.startRule(token.value))
